+++ src/Options/Error-handling/errors.js
@@ -1,20 +1,21 @@
-export class KongouError extends Error {
+class KongouError extends Error {
   constructor(code, message) {
     super(`${message} [${code}]`);
     this.name = "KongouError";
   }
 }
 
-export class KongouClientError extends KongouError {
+class KongouClientError extends KongouError {
   constructor(code, message) {
     super(code, message);
     this.name = "KongouClientError";
   }
 }
 
-export class KongouServerError extends KongouError {
+class KongouServerError extends KongouError {
   constructor(code, message) {
     super(code, message);
     this.name = "KongouServerError";
   }
 }
+module.exports = { KongouError, KongouClientError, KongouServerError };